Allow credentials to be passed multiple ways.
