+++ tests/utils.js
@@ -1,0 +1,11 @@
+
+function assertNoError(e) {
+  // Assertion to ensure that error is omitted inside a callback.
+  if (e) {
+    throw e;
+  }
+}
+
+module.exports = {
+  assertNoError,
+};
